refactor: rename provider directory to providers and relocate AppProviders
