+++ src/entities/user/api/update-user-profile.ts
@@ -4,7 +4,7 @@ import { USER_QUERY_KEY } from './queryKey'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { USER_API } from './endpoint'
 import useUserStore from '@/src/shared/store/userStore'
-import { useAuth } from '@/src/shared/provider'
+import { useAuth } from '@/src/shared/providers'
 
 export const patchUser = async (data: UpdateUserType) => {
   try {
@@ -19,7 +19,7 @@ export const patchUser = async (data: UpdateUserType) => {
 export const useUpdateUser = () => {
   const queryClient = useQueryClient()
   const updateStateUser = useUserStore((state) => state.updateStateUser)
-  const {token} = useAuth()
+  const { token } = useAuth()
   return useMutation({
     mutationFn: (data: UpdateUserType) => patchUser(data),
     onSuccess: (_, data) => {
@@ -28,8 +28,10 @@ export const useUpdateUser = () => {
         profile_url: '',
         description: '',
       })
-      if (token){
-        queryClient.invalidateQueries({ queryKey: USER_QUERY_KEY.myProfile(token) })
+      if (token) {
+        queryClient.invalidateQueries({
+          queryKey: USER_QUERY_KEY.myProfile(token),
+        })
       }
     },
   })
