+++ src/features/modal/ShareModal.tsx
@@ -5,7 +5,7 @@ import { Modal } from '@/src/shared/ui'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
 import { PlanType } from '@/src/entities/plan'
-import { useAuth, useToast } from '@/src/shared/provider'
+import { useAuth, useToast } from '@/src/shared/providers'
 import close from '@/src/assets/icon/medium/cross.svg'
 import link from '@/src/assets/icon/medium/link.svg'
 import share_white from '@/src/assets/icon/medium/share_white.svg'
