+++ src/shared/providers/FcmProvider.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect } from 'react'
 import { initFCM } from '@/src/features'
-import { useAuth } from '@/src/shared/provider'
+import { useAuth } from '@/src/shared/providers'
 
 export const FcmProvider = () => {
   const { token } = useAuth()
